got the sigmoid normalizer working.  some refactoring was done on structure.  see the SOM test and the SigmoidNormalizer test for reference

+++ work/tbeck/somtk/normalizers/sigmoidnormalizer.h
@@ -33,8 +33,8 @@ private:
     /// The epsilon value to use for the sigmoid squashing
     double epsilon;
 
-    /// The number of standard deviations fro which to calibrate the sigmoid curve
-    int sigmaStep;
+    /// The number of standard deviations for which to calibrate the sigmoid curve
+    double sigmaStep;
